content: generateAlbumChronologyLinks
